update records with more information
also a bunch of other small things

+++ pool/reward_record.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass
 
 from chia.types.blockchain_format.sized_bytes import bytes32
-from chia.util.ints import uint64
+from chia.util.ints import uint64, uint32
 from chia.util.streamable import streamable, Streamable
 
 
@@ -10,6 +10,6 @@ from chia.util.streamable import streamable, Streamable
 class RewardRecord(Streamable):
     launcher_id: bytes32  # This uniquely identifies the singleton on the blockchain (ID for this farmer)
     claimable: uint64  # The amount of token claimable
-    block_height: uint64  # The height of the block to claim
+    block_height: uint32  # The height of the block to claim
     coins_hash: bytes32  # The coin's identifier
     timestamp: uint64  # The timestamp of the reward
